started reconstructing the api

+++ OsmiumGL/API/include/RenderedObject.h
@@ -14,7 +14,7 @@ namespace std {
 
 using MaterialHandle = unsigned long;
 using MatInstanceHandle = unsigned long;
-using MeshHandle = unsigned long;
+using MeshHandle = unsigned int;
 using TextureHandle = unsigned long;
 struct RenderedObject {//Struct entries are ordered like they are in the render tree
     MaterialHandle material;//this is mostly the pipeline
